test(plugins): split remote AWS wrapper coverage
Resolve the max-lines review without suppressing the repository rule, preserving the wrapper regression assertions in a focused test module.

Work-Item: CodySwannGT/lisa#1706

Work-Item: CodySwannGT/lisa#1706
