+++ tests/unit/strategies/remote-agent-aws-setup.test.ts
@@ -1,4 +1,3 @@
-/* eslint-disable max-lines -- Remote AWS strategy coverage keeps installer, wrapper, and bootstrap contracts together. */
 /**
  * Vendor-neutral remote AWS bootstrap and platform-adapter coverage.
  * @module tests/unit/strategies/remote-agent-aws-setup
@@ -21,8 +20,6 @@ import { afterEach, describe, expect, it } from "vitest";
 import { installRemoteAgentAws } from "../../../plugins/src/base/scripts/install-remote-agent-aws.mjs";
 
 const temporaryDirectories: string[] = [];
-const REMOTE_SETUP_WRAPPER_PATH =
-  "all/create-only/scripts/remote-agent-aws-setup.sh";
 const REMOTE_SETUP_SCRIPT_PATH =
   "plugins/src/base/scripts/remote-agent-aws-setup.sh";
 const CURSOR_ENVIRONMENT_PATH = ".cursor/environment.json";
@@ -224,66 +221,6 @@ describe("remote AWS runtime parity", () => {
   });
 });
 
-describe("remote AWS setup wrapper", () => {
-  it("prints installation guidance when npm is unavailable", () => {
-    const emptyPath = temporaryDirectory();
-    const result = spawnSync("/bin/bash", [REMOTE_SETUP_WRAPPER_PATH], {
-      cwd: path.resolve("."),
-      encoding: "utf8",
-      env: { ...process.env, PATH: emptyPath },
-    });
-
-    expect(result.status).toBe(1);
-    expect(result.stdout).toBe("");
-    expect(result.stderr).toBe(
-      "remote-agent-aws-setup: install @codyswann/lisa before running this setup script\n"
-    );
-  });
-
-  it("delegates to the installed Lisa script with every argument unchanged", () => {
-    const root = temporaryDirectory();
-    const binaryDirectory = path.join(root, "bin");
-    const packageRoot = path.join(root, "node_modules");
-    const installedScript = path.join(
-      packageRoot,
-      "@codyswann/lisa/plugins/lisa/scripts/remote-agent-aws-setup.sh"
-    );
-    const argumentLog = path.join(root, "arguments.log");
-    mkdirSync(binaryDirectory, { recursive: true });
-    mkdirSync(path.dirname(installedScript), { recursive: true });
-    writeFileSync(
-      path.join(binaryDirectory, "npm"),
-      `#!/bin/bash\nprintf '%s\\n' "${packageRoot}"\n`
-    );
-    writeFileSync(
-      installedScript,
-      '#!/bin/bash\nprintf \'%s\\0\' "$@" > "$FAKE_ARGUMENT_LOG"\nexit 23\n'
-    );
-    chmodSync(path.join(binaryDirectory, "npm"), 0o700);
-    chmodSync(installedScript, 0o700);
-    const argumentsToForward = ["plain", "two words", "--flag=value"];
-
-    const result = spawnSync(
-      "/bin/bash",
-      [REMOTE_SETUP_WRAPPER_PATH, ...argumentsToForward],
-      {
-        cwd: path.resolve("."),
-        encoding: "utf8",
-        env: {
-          ...process.env,
-          PATH: binaryDirectory,
-          FAKE_ARGUMENT_LOG: argumentLog,
-        },
-      }
-    );
-
-    expect(result.status).toBe(23);
-    expect(readFileSync(argumentLog, "utf8").split("\0").slice(0, -1)).toEqual(
-      argumentsToForward
-    );
-  });
-});
-
 /* eslint-disable sonarjs/no-os-command-from-path -- Test-only PATH shims inject the fake aws executable. */
 describe("remote AWS bootstrap script", () => {
   /**
@@ -388,4 +325,3 @@ fi
   });
 });
 /* eslint-enable sonarjs/no-os-command-from-path -- End test-only PATH shim scope. */
-/* eslint-enable max-lines -- End remote AWS strategy coverage. */
